9章のコードを追加

+++ app/src/Controller/AnswersController.php
@@ -39,7 +39,7 @@ class AnswersController extends AppController
             return $this->redirect(['controller' => 'Questions', 'action' => 'view', $answer->question_id]);
         }
 
-        $answer->user_id = 1; // @TODO ユーザー管理機能実装時に修正する
+        $answer->user_id = $this->Auth->user('id');
         if ($this->Answers->save($answer)) {
             $this->Flash->success('回答を投稿しました');
         } else {
@@ -59,7 +59,10 @@ class AnswersController extends AppController
     {
         $answer = $this->Answers->get($id);
         $questionId = $answer->question_id;
-        // @TODO 回答を削除出来るのは回答投稿者のみとする
+        if ($answer->user_id !== $this->Auth->user('id')) {
+            $this->Flash->error('他のユーザーの回答を削除することはできません');
+            return $this->redirect(['controller' => 'Questions', 'action' => 'view', $questionId]);
+        }
 
         if ($this->Answers->delete($answer)) {
             $this->Flash->success('回答を削除しました');
